test: move ALS keep-alive test to its own file
node-http.test.ts has a pre-existing unrelated ECONNREFUSED failure
(request via http proxy, issue#4295) that masks a clean fail-before/
pass-after signal; keep the regression test in a standalone file.
