+++ test/js/node/http/node-http.test.ts
@@ -7,7 +7,6 @@
  */
 import { bunEnv, bunExe, exampleSite, randomPort, tls as tlsCert } from "harness";
 import { createTest } from "node-harness";
-import { AsyncLocalStorage } from "node:async_hooks";
 import { EventEmitter, once } from "node:events";
 import nodefs from "node:fs";
 import http, {
@@ -3749,57 +3748,4 @@ it("OutgoingMessage outputData is per-instance and _flushOutput is defined", () 
   const d = new OutgoingMessage();
   c.outputData.push({ data: "y", encoding: "utf8", callback: null });
   expect(d.outputData.length).toBe(0);
-});
-
-it("AsyncLocalStorage context is preserved across keep-alive socket reuse", async () => {
-  const als = new AsyncLocalStorage<string>();
-  const server = createServer((req, res) => res.end("ok"));
-  server.listen(0, "127.0.0.1");
-  await once(server, "listening");
-  const { port } = server.address() as AddressInfo;
-  const agent = new Agent({ keepAlive: true, maxSockets: 1 });
-  const sockets: unknown[] = [];
-  const seen: Record<string, Record<string, string | undefined>> = {};
-
-  try {
-    const one = (store: string) =>
-      new Promise<void>((resolve, reject) =>
-        als.run(store, () => {
-          const req = http.request({ host: "127.0.0.1", port, path: "/", agent }, res => {
-            seen[store].response = als.getStore();
-            res.on("data", () => (seen[store].data = als.getStore()));
-            res.on("end", () => {
-              seen[store].end = als.getStore();
-              resolve();
-            });
-            res.on("error", reject);
-          });
-          seen[store] = {};
-          req.on("socket", s => {
-            sockets.push(s);
-            seen[store].socket = als.getStore();
-          });
-          req.on("error", reject);
-          req.end();
-        }),
-      );
-
-    await one("R1");
-    await new Promise<void>(r => setImmediate(() => setImmediate(r)));
-    await one("R2");
-    await one("R3");
-
-    // The same pooled socket served all three requests.
-    expect(sockets).toEqual([sockets[0], sockets[0], sockets[0]]);
-    // Each request's 'socket'/'response'/'data'/'end' saw its own store,
-    // not the context the native socket captured at connect time (R1).
-    expect(seen).toEqual({
-      R1: { socket: "R1", response: "R1", data: "R1", end: "R1" },
-      R2: { socket: "R2", response: "R2", data: "R2", end: "R2" },
-      R3: { socket: "R3", response: "R3", data: "R3", end: "R3" },
-    });
-  } finally {
-    agent.destroy();
-    server.close();
-  }
 });
